requirements.txt
bluepy
futures
blue-st-sdk
paho-mqtt
python-dotenv